refactor: move languages dropdown to separated component
