+++ src/components/shared/header/header.jsx
@@ -1,17 +1,13 @@
 import { Popover, Transition } from '@headlessui/react';
 import { MenuIcon, XIcon } from '@heroicons/react/outline';
-import { ChevronDownIcon } from '@heroicons/react/solid';
 import React, { Fragment } from 'react';
 
-import Button from 'components/shared/button';
 import Container from 'components/shared/container';
 import Link from 'components/shared/link';
 import CiliumLogo from 'images/cilium-logo.inline.svg';
 import Logo from 'images/logo.inline.svg';
 
-import EnglishFlag from './icons/english.inline.svg';
-import FrenchFlag from './icons/french.inline.svg';
-import ChineseFlag from './icons/сhinese.inline.svg';
+import Languages from './languages';
 import SubMenu from './sub-menu';
 
 const navigation = [
@@ -60,17 +56,7 @@ const navigation = [
   },
 ];
 
-const languages = [
-  { icon: EnglishFlag, name: 'English' },
-  { icon: FrenchFlag, name: 'Français' },
-  { icon: ChineseFlag, name: '简体中文' },
-];
-
-function classNames(...classes) {
-  return classes.filter(Boolean).join(' ');
-}
-
-const Header = () => (
+const Header = ({ switchLanguages }) => (
   <Popover>
     {({ open }) => (
       <>
@@ -90,7 +76,7 @@ const Header = () => (
                 </div>
               </div>
             </div>
-            <div className="flex space-x-10 items-center lg:hidden lg:space-x-3.5">
+            <div className="flex items-center space-x-10 xl:space-x-8 lg:hidden">
               {navigation.map(({ name, href, menuItems }) => {
                 const hasSubmenu = menuItems?.length > 0;
                 return (
@@ -110,68 +96,14 @@ const Header = () => (
               })}
             </div>
             <div className="relative lg:hidden">
-              <div className="absolute -translate-y-1/2 top-1/2 -left-36">
-                <Popover className="relative">
-                  {({ open }) => (
-                    <>
-                      <Popover.Button
-                        className={classNames(
-                          open ? 'text-gray-900' : 'text-black',
-                          'group bg-white rounded-md flex items-center text-base focus:outline-none font-bold leading-none transition-colors duration-200 hover:text-gray-1'
-                        )}
-                      >
-                        <EnglishFlag className="mr-3" />
-                        <span>English</span>
-                        <ChevronDownIcon
-                          className={classNames(
-                            open ? 'text-gray-900' : 'text-black',
-                            'h-5 w-5 group-hover:text-gray-1 transition-colors duration-200'
-                          )}
-                          aria-hidden="true"
-                        />
-                      </Popover.Button>
-
-                      <Transition
-                        show={open}
-                        as={Fragment}
-                        enter="transition ease-out duration-200"
-                        enterFrom="opacity-0 translate-y-1"
-                        enterTo="opacity-100 translate-y-0"
-                        leave="transition ease-in duration-150"
-                        leaveFrom="opacity-100 translate-y-0"
-                        leaveTo="opacity-0 translate-y-1"
-                      >
-                        <Popover.Panel
-                          className="absolute z-10 w-screen max-w-[180px] px-2 mt-3 transform -translate-x-1/2 left-1/2 sm:px-0"
-                          static
-                        >
-                          <div className="overflow-hidden rounded-lg shadow-lg ring-1 ring-gray-900 ring-opacity-5">
-                            <div className="relative grid gap-6 px-5 py-6 bg-white sm:gap-8 sm:p-8">
-                              {languages.map(({ icon: Icon, name }) => (
-                                <button
-                                  type="button"
-                                  key={name}
-                                  className="flex items-center px-3 py-2 -m-3 space-x-5 text-base font-medium text-gray-900 rounded-md hover:bg-gray-50"
-                                >
-                                  <Icon />
-                                  <span>{name}</span>
-                                </button>
-                              ))}
-                            </div>
-                          </div>
-                        </Popover.Panel>
-                      </Transition>
-                    </>
-                  )}
-                </Popover>
-              </div>
+              {switchLanguages && <Languages />}
               <Link
-                className="block pl-8 border-l-2 border-gray-2"
+                className="block"
                 to="https://cilium.io/"
                 target="_blank"
                 rel="noopener noreferrer"
               >
-                <CiliumLogo />
+                <CiliumLogo className="xl:h-7 xl:w-auto" />
               </Link>
             </div>
           </nav>
