+++ src/layouts/main/main.jsx
@@ -4,9 +4,9 @@ import React from 'react';
 import Footer from 'components/shared/footer';
 import Header from 'components/shared/header';
 
-const MainLayout = ({ children }) => (
+const MainLayout = ({ children, switchLanguages }) => (
   <>
-    <Header />
+    <Header switchLanguages={switchLanguages} />
     <main>{children}</main>
     <Footer />
   </>
@@ -14,6 +14,11 @@ const MainLayout = ({ children }) => (
 
 MainLayout.propTypes = {
   children: PropTypes.node.isRequired,
+  switchLanguages: PropTypes.bool,
+};
+
+MainLayout.defaultProps = {
+  switchLanguages: false,
 };
 
 export default MainLayout;
